complete restyling of test/benchmark code to use assert/require
